Add files via upload
With this project I was asked to collect or recreate data of my choice by collecting at least 50 observations.The observations you collect may be discrete data or continuous data but cannot be categorical in nature.  I can generate or collect the observations myself, but I may not use mass emailing to the class to do so.   I can obtain the data from internet sites, but I must include where you collected the data from as a reference. determine a null and alternative hypothesis on the population mean of the data.  I can select the values testing on arbitrarily, or I can select values based on some previous knowledge.  direction of the alternative hypothesis is left up to me. Documented findings per the organization provided at the end.